Server/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.4
psycopg[binary]==3.2.9
python-multipart==0.0.20
sqlalchemy==2.0.36
alembic==1.14.1
python-dotenv==1.0.1
pytest==8.2.2
pytest-asyncio==0.24.0
httpx==0.27.0 